[FIX] *: set ondelete policy of required Selection fields
With this commit, Selection fields with `required=True` which are
extended via `selection_add` are given proper ondelete policies to
ensure the cleanup of records containing these extended options during
uninstall of the extending module.

This commit also gets rid of leftover uninstall hooks that would handle
the same set of problems prior to the implementation of the ondelete
mechanism in Selection fields.

closes odoo/enterprise#9117

Related: odoo/odoo#46325
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>