[FIX] currency_rate_live: update only available rates
When requesting rates from banxico for a given date range, it happens
that not all requested rates are available at that particular moment.
In that case, the parse_function fails.

With this commit, if one of the currency rate is missing, not a single
rate is created and a warning is logged. That way, the next cron
execution may update the rates if they are available and no partial data
are written.

Also, the test is verify that a rate was created only if a result was
retrieved for banxico. Instead of a failure, the warning will make the
runbot tests yellow.

closes odoo/enterprise#8788

X-original-commit: 795aeaee63f9bc023a8fa7e7d9cc0074b682ee48
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>