[FIX] sign: don't freak out users about fake integriy issue
Since odoo/odoo@0305013 and odoo/enterprise@d1bd484, compute_sudo is
False by default for non-stored fields.

The integrity field of a signature request depends on fields that may
not be accessible to the current user (geolocation, ip address, etc. of
sign events) but that are required to check the integrity of the
document's history.

This means that if this field is not computed in a sudoed environment,
any user without sufficient access rights (notably the group 'Sign
Manager') will have a rather worrying 'the integrity of this document
could not be verified', implying that some shady black hat hackers could
have modified their db while in fact it's just an access rights
restriction.

This commit set the integrity field to be computed in a sudoed
environement to avoid this issue.

X-original-commit: 438612a341d95f924a39aaa9b1cfa39ae08b2c48