[FIX] sign: do not leak private information in the activity report
Depending on who's the last signer, the report may include information
that should normally be private (IP address of signers, geolocation).
This happened because the report was printed using the environment of
the current user, which may have the group 'Sign Manager' - if so, then
extra information is included in the report and is then sent to all
signers.

This commit mitigate this problem by using the public user to generate
the report. Note that this assumes that:
- the public user does not have the group 'Sign Manager'
- the public user exists

Since having any of these assumption not met would cause way, _way_ more
serious problems (the leaked data would be more substantial in the first
case, everything would probably be broken in the second), this seems
like a reasonable tradeoff.

X-original-commit: 4b98c09bc8be25451fcbdb624a57c751d3366047