[FIX] planning: Fix read group employee with date
Before this commit:
when we have 'planning_expand_employee', they are a special read
group to get all employees having a slot in the past and next 30
days. But if we select a filter with start_datetime (for example
'future'), we won't search employees having a slot around the
calendar week but having a slot around today.

After this commit:
We search employees having a slot around the calendar date even
a filter is selected.

closes odoo/enterprise#8761

X-original-commit: 7cf7f47e68f9bf0942e5240973fafe16cec02bfa
Signed-off-by: jbm-odoo <jbm-odoo@users.noreply.github.com>