[FIX] web: grouped kanban card doesn't have right size on mobile
Before this commit the grouped kanban card with a search panel didn't
have the right width because they used the size of the kanban tabs.

After this commit, on mobile, the size used is the full width of
the screen like the ungrouped kanban.

Steps to reproduce:

Go in an app with a search panel (e.g. employees hr)
Use the filter to group the kanban view (e.g. Department) => BUG
Task ID : 2200099

closes odoo/enterprise#8811

Forward-port-of: odoo/odoo#46178
X-original-commit: 361233c36bd501998ecfc88c1df67ecc0022bc79
Signed-off-by: rfr-odoo <rfr-odoo@users.noreply.github.com>