[FIX] web_enterprise: Implement mobile company switcher
Commit odoo/odoo@a5b6f31 introduced a new company switcher widget but forgot to implement.
the mobile part.

This commit fixes this oversight.

To avoid any breaking changes with possible customization, this fix has a
minimal diff to implement it, but no the cleanest.

The mobile widget and its integration with the desktop one is improved
in master with odoo/odoo#46690 and odoo/enterprise#8965

Task 2205603

closes odoo/enterprise#9373

X-original-commit: cff73ea89cb16668b5402300bdf7150e25535f34
Related: odoo/odoo#48045
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: lul-odoo <LucasLefevre@users.noreply.github.com>