[FIX] stock_barcode: scan a package on a partially available line
Let's say pack1 contains 10 product1.

If a move of 10 product1 is created, it is fully assigned and the
package_id and result_package_id are already set to pack1. Scanning
pack1 in the client action will only set the qty_done, the entire
package will be moved.

If a move of 1 product1 is created, it is fully assigned but only
package_id is set, not result_package_id since the package won't
be entirely moved to the destination location.

As we always try to move the entire package when scanning the pack, we
have to always set result_package_id, else validating the move will took
all product1 out of pack1 and leave an empty pack1 at the osurce
location

opw-2198530

closes odoo/enterprise#8938

X-original-commit: 8589bc20a42bae64997167add4ca81a4ba6b53b4
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>