[FIX] web_enterprise, mail_enterprise: scroll issue after swipe aborting
(1) Before this commit, when a user wanted to cancel the trigger of the
action by swipping in the opposite direction, the item's background remains
gray and the parent wasn't scrollable anymore.
Indeed, The DOM was not restored if you swipped in the opposite way before
the second threshold; It worked if you did it after this boundary.
Return was called preventing the DOM to be fixed.

Now the DOM is restored as it should be but we had to "reorganise" (/rewrite
is probably a word to proscribe as this is a fix) because specs changed
and it will be easier to add future fixes (if needed) in a more concise
source code.

(2) We also add a small improvement to progressively display the action icon
when you are swipping. By doing this, you know when you must release your
finger to be able to execute the action. We also fix the icon alignement.

Steps to reproduce (for point 1):
* On mobile, open the SysTray Messaging
* Then swipe on a thread that have an unread message without releasing the touch
* Swipe in the opposite before the threshold (< 15px)
* Then release the touch
* Now the parent element of initial touch is not scrollable anymore => BUG

Task ID : 2198394

closes odoo/enterprise#9006

X-original-commit: 7633a58da939eff9639697375a157632fa8f4938
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>