[FIX] industry_fsm(_sale, _report): Split and improve the onboarding tour.
Currently the FSM tour is only appear when the industry_fsm_sale is
installed and structure of the tour is not dependency compatible.
Tour of industry_fsm_sale is loaded inside the industry_fsm_report
but when you install the report module then fsm_sale module is
not installed.

So in this commit, we split the tour into the two module. main tour
will be inside the industry_fsm. From industry_fsm_sale and
industry_fsm_report we have added the extra steps to the main tour.
Also improved steps and description of the steps for the tour.
We also added the product creation steps as we need a product from
fsm_sale module while the tour is started from industry_fsm.

closes odoo/enterprise#8939

Taskid: 2188560
Closes: #8939
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>