[IMP] hr_payroll: set boolean_toggle on 'active' field on form view
Currently, to activate/deactivate records with the 'active' checkbox
user has to switch to edit mode of the form.
So the purpose of the task is to allow the user to activate/deactivate
records from the readonly mode of the form view.

In this commit, we set widget='boolean_toggle' on the 'active' field in form
view.

TaskID : 2206794
Related: https://github.com/odoo/odoo/pull/46567
Closes : #8918

Related: odoo/odoo#46567
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>