[FIX] web_enterprise: Call on_attach_callback on systray widgets
Currently, the method `on_attach_callback` is never called on any Widget in the
systray.

Calling `on_attach_callback` is required if any systray widget uses an OWL
component via the ComponentWrapper. Without the callback, the component is never
considered mounted to the DOM and no event handler is triggered.

The Community menu test cannot run since the menu is replaced by the Enterprise
menu. Hence the Community test is replaced by the Enterprise one (in the same
way the menu is replaced).

closes odoo/enterprise#8887

Related: odoo/odoo#46566
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>