[FIX]: helpdesk_timesheet: allow_timesheet_timer wrongly computed
The _compute_allow_timesheet_timer method completely overrides the parent method,
without keeping the expected behavior.

Now, a timesheet timer is allowed if set in the helpdesk team OR if timesheets are
allowed for associated project.

New tests added in helpdesk_timesheet to be sure to keep this behavior after future
modifications.

closes odoo/enterprise#8923

Taskid: 2206820
Related: odoo/odoo#46692
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>