[FIX] web_mobile: phone vibrates when Odoo app is minimized
Before this commit, the phone vibrated and a notification was
displayed in case of connection loss/restore.

This vibration seems annoying when the Odoo app is on background
as we don't know what's happening.

This is why we decided to remove it. The text notification inside
the webview is enough.

Task ID : 2207465

closes odoo/enterprise#10771

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>