[IMP] sale_subscription: a subscription product cannot be storable
As subscriptions are not integrated in the inventory, it should not
be possible to set a subscription product as a storable product.

To prevent that:
- if the product is a subscription product, when 'storable' product
is selected, a warning message is displayed and the product type
is reset to False.
- if the product is a storable product, when 'subscription product'
is checked, a warning message is displayed and the field is reset
to False.

A constraint on the 'product.template' model has also been added to
guarantee that.

closes odoo/enterprise#8739

Taskid: 2158101
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>