[FIX] project_forecast: Manage slots linked to private project
After this commit:
Planning manager can see all shifts (even if linked to private project)
Planning user can only see shifts linked to project that he can see
(if project is private and he is not follower of this project, he
can't see this shift even he is assigned on this shift.)
Works in back end and in portal.
Show warning message in slot if the project is invisible for
selected employee. And remove button to send him this slot.

Task ID 2202941

closes odoo/enterprise#9359

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>