[IMP] planning: easily modify shifts with templates
=======
Purpose
=======

Allow to easily modify a shift using templates (e.g. switch from an
afternoon shift to a morning shift in one click).

===============
Specifications
===============

Always display the 'Shift Templates' field on shifts - If a shift was
created using a template, the corresponding template should be displayed
as being selected - If the user selects another template, apply an
onchange on the role, project, task and date fields (if set on the
template) - If the user unselects the template, empty the role, project,
task fields

closes odoo/enterprise#9443

Taskid: 2208080
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>