[IMP] l10n_be_hr_payroll: Manage whole 'Credit Time' mechanism
Purpose
=======

It is possible in Belgium to take a credit time full time and so we have to support this.
Furthermore, the credit time days appears on the payslip (at 0 € but it must be
visible in the payslip)

Other issues are occuring when computing unclassical payslips. All the improvements/fixes that
have been made are listed below.

Specification
=============

1/ Manage out of contract time.

When computing a payslip, we don't take into account the fact that the contract might start
after the 1rst of the month or end before the last day of the month.

Currently, even if the employee worked 2 days on the month, the payslip is computed with
the full wage.

Introduce a new work.entry.type 'Out of Contract'. When this is the case, a worked_days_line
is generated, counting the corresponding hours/days. And this will also reduce the wage on the
payslip to take only the worked hours into account.

2/ Adapt several hr_payroll/hr_work_entry/... methods to allow inheritence on l10n_be_hr_payroll

3/ Adapt all the salary rules for the CP200 regular pay.

Indeed, if the basic salary is equal to 0 (could happen on a full-time credit time calendar),
then it makes no sense to compute internet/mobile/car benefits in kind, for example.

No need to compute the withholding taxes too, ...

4/ Introduce a new work entry type 'Credit Time'

This will be used to generate work entries on the employee schedule, and will be reported
on the payslip too.

A structure type (eg: CP200) can configure a work entry type to use when generating the work
entries.

5/ Add the full-time credit time options.

It's now possible in Belgium to take a career break without working at all. Add this option.

6/ Fix meal vouchers computation

The meal vouchers are counter over the month for a regular payslip, without taking the
contract boundaries into account. No need to give 22 meal vouches if the employee started
at the end of the month.

7/ Fix the basic wage computation.

If the employee doesn't work, the basic wage computation shouldn't crash. Take this edge
case into account. Same for the special social cotisations and the employement bonus.

8/ Add tests

Ensure the Credit Time mechanism by writing 2 tests (Full time + 4/5 credit time).

Those tests use a complete employment contract with all the possible advantages in
Belgium, check the work entries generation and the payslips computation to ensure that
all the lines for both old and new contracts are correct.

closes odoo/enterprise#7173

Taskid: 2145813
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: laa <laa@odoo.com>