[FIX] documents_account: expands coverage of settings to all moves.
Before this commit, it was possible to link folders to accounting
journal that were related to `account.moves` that are not invoices.

This commit expands the coverage of the journal linking settings
to all relevant `account.move`.

Additionally a domain to the journal is added excluding types of
`bank` and `cash` because their attachment is not directly attached to
their record.

task-2146838

closes odoo/enterprise#10190

X-original-commit: ff960131c419a0982c62f998c71cd4f528f93b9e
Signed-off-by: Thanh Dodeur <ThanhDodeurOdoo@users.noreply.github.com>