[ADD] event_social: add the communication support
Purpose
=======
We want to be able to post on social medias in the event communication
tab.

Specifications
==============
For that purpose, we added the social post template to the event mail
and to the event type mail, so the user can choose a social template
that will be sent on the scheduled date.

To be coherent with the mail/SMS template, the social post is sent in
the event CRON. We create a social post with the value of the social
post template and with the social accounts linked to the event mail.

Links
=====
Task-2127615
See odoo/odoo/pull/46304
See odoo/enterprise/pull/7701
See odoo/upgrade/pull/848