[REM] event_social_facebook: drop the module
The endpoint we used to post on the Facebook events will be depreciate
on the 2020-02-03. So, this module become useless and need to be dropped.
> `https://developers.facebook.com/docs/graph-api/changelog/2020-02-03-endpoint-deprecations`

We are working on a new way to post on social accounts in event communication
tab. This will be done in another task, targeting master.

This commit basically reverts 2f17fdc2756be1eadd0cc53aa0ab3986c829f5b8 (coming from odoo/enterprise#8058) . Came in 13.2, died in 13.2.

Task 2210120

closes odoo/enterprise#9441

X-original-commit: ee73ddd27ae6497f3c93416bb4d8acef05800b17
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>