[FIX] account_followup: join invoices no duplicate
When a report is generated, the header/foot/bodies are split before
being given to wkhtmltopdf in ir.actions.report `_prepare_html`.

This can be done because we have special class .header for header,
.footer for footer and .article for bodies.

But in followup report, if the option "Join open Invoices"
(join_invoices) is enabled, we would get a structure such as:

```
<div class="header"></div>
<div class="article">
    <div class="page">body of follow report</div>
    <div class="page">
        <div class="header"></div>
        <div class="article">body of first invoice</div>
        <div class="footer"></div>
    </div>
    <div class="page">
        <div class="header"></div>
        <div class="article">body of second invoice</div>
        <div class="footer"></div>
    </div>
</div>
<div class="header"></div>
```

This cause an issue, because then when we print the 3 bodies, we will
print the first one that also contain the second and third bodies. But
then the second and third bodies are also printed and we get duplicates,
so we get these bodies:

body1:
- body of follow report
- header first invoice
- body of first invoice
- footer first invoice
- header second invoice
- body of second invoice
- footer second invoice
body2:
- body of first invoice
body3:
- body of second invoice

opw-2207706
closes #9072

closes odoo/enterprise#9094

X-original-commit: 8b348e893d5d8514ba69cf70f0438328aba67f8a
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>