[FIX] stock_barcode: renames test's shelves
As we usually use WH/Stock as location in test, we also have it's child
locations (Shelf 1, Shelf 2, ...). But, in some test, we created also
locations we name also Shelf 1 and Shelf 2.
Because of that, it can happen in test tour that the tour doesn't select
the same location (for instance, want to select Shelf 1, but hey, we
have two Shelf 1 !). The result could be we have during the tour an
another page with the same name than an existing page.

So, to avoid this undeterminist behavior, this commit renames all
locations created in tests with a "Shelf n" name into "Section n".

closes odoo/enterprise#9144

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>