[FIX] project_forecast: Keep default project on shift
Before this commit:
If we have a default project in context (for example, when we click
on 'planning' button on the kanban view of project), and we create
a new shift, the default project is present in context but
there are no default task so the project will be erase.

After this commit:
Even if there are no task on a new shift, the default project value
is kept.

closes odoo/enterprise#9139

X-original-commit: be08987fd00696c257bc44d7082bdedb575b64a7
Signed-off-by: jbm-odoo <jbm-odoo@users.noreply.github.com>