[FIX] l10n_co_edi_ubl_2_1: also put 0% tax info in e-invoice xml
We need to add information in the xml about which
taxes are applied with their edi code.
Before, we searched for these taxes by taking the
tax lines of the invoice (tax_line_id).  However,
there is no tax_line_id generated for the 0% taxes.

Now, in order to have all the taxes, they are put
anyways on the base amounts, where we have the tax_ids.

Task id: 2209925

closes odoo/enterprise#9176

X-original-commit: ba26364397045a288e016b886082a5f9af908c09
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>