[FIX] account_followup: fix Followup Report Letter
Install Germany - Accounting.
In General Settings, set "European A4 for DIN" as the Format and
"external_layout_din5008" as the layout.
Print Followup report letter

The report has misplaced elements, the other company address should be
present on the left and there is the extra address below the FollowUp
header.
Modifying the style to hide the address below the followup title
and the xml addon for din to fetch the address from the partner fix the
issue

Companion patch of https://github.com/odoo/odoo/pull/46443

opw-2187240

closes odoo/enterprise#9377

X-original-commit: a0e5ae67e1e4c775da8a6c25b96853319edcb119
Related: odoo/odoo#48057
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>