[IMP] sale: ignore null SO in reconciliation matching
It is unlikely that anyone would make an empty transfer for a SO with
zero amount. By ignoring these lines, we save on the costly POSITION
searches. For example, the website_sale_event App can lead to a high
number of SO with a zero amount (registrations for free events).

X-original-commit: f6b881f6bee4ee8cc0b742b6a004f6078db9907f