[FIX] account_invoice_extract: don't set a field as required when not necessary
Adding required fields on company model is annoying for tests. Indeed multi
company tests may run during module installation, notably in CRM. Those tests
generally require to create a new company which is not possible if the column
is required but module not already loaded in registry. This is the case
if for example account_invoice_extract is installed and updating CRM
when testing website_crm_score.

When possible, avoid required fields on a such used model as company. In this
case moreover setting a boolean as required makes no sense and has no impact
in code.

closes odoo/enterprise#9381

Related: odoo/odoo#48059
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>