[FIX] social_linkedin: fix crash during post sending
Bug
===
If we send an image on LinkedIn, we need to upload it first. If the
image upload fail, we raise an error. But, if the post is sent in the
social CRON (if the post is scheduled) and if the image upload failed,
then the error will be raised, restarting the CRON, making it do an
infinite loop.

Solution
========
Catch the error in the social CRON and mark the post sending as failed.

Task-2206477
PR #8857

closes odoo/enterprise#9313

X-original-commit: 31561ec1b4fcb013e0a9d60e44b15bce55f7235a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: std-odoo <std-odoo@users.noreply.github.com>