[FIX] account: "load more" button in manual reconciliation
Status : For the moment, the "load more" button skip some lines each time we use it.
The splice is not used correctly. For example, on a array of 75, we see the first
30 lines. The splice function cuts the array in two : The first 30 lines on one and last
45 lines in an other. After the cut, we say that an index is at 30 to split on the load more.
When we click on "Load more", we check the second array. But, as we use the index at 30 for
the splice function, we skip the first 30 lines of the second array (with only 45 lines so we
needed to take the first 30 lines) and only show the last 15 lines.

Expected : When we click on "load more" button, we need to respect the pagerIndex
and not remove elements.

Fix : Changing the way we use the splice. Using slice instead of splice.

note: enterprise forward port of community odoo/odoo#47803

opw-2171734
closes #9305

closes odoo/enterprise#9318

X-original-commit: c617fdb891aaaa78eba1987feb8cc5ad17f2af99
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>