[IMP] web_enterprise: apply more precedence for mobile search css rules
Before this commit, few of the css rules related to mobile searchview were
having less precedence and so UI was not as expected.

For example, the css for the "See results" CTA were not applied correctly (e.g.
`font-size`).

This commit fixes this issue by adding proper precedence for the mobile
searchview related rules and thus avoiding such issues in future.

Task ID: 2152164

closes odoo/enterprise#7274

Related: odoo/odoo#41898
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>
Co-authored-by: Dharmraj Jhala <dja@odoo.com>