[REF] mail_enterprise: update discuss helpers
In community, the 'createDisuss' helper has been updated to use 'dom.prepend' to
benefit of its 'on_(attach|detach)_callback' calls. This allows the mail test suite
environment to be closer to the actual action behaviour (having those methods called
each time it is created/destroyed).

This is why one of the mail tests which asserted specifically that the action would
not crash when using these methods has been removed.

closes odoo/enterprise#9328

Related: odoo/odoo#47865
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>