[FIX] voip: no contact shown on incoming call
Before this commit, for an incoming call, only the phone number was
shown, even if the contact is registered as a contact.

The phone number used to search the contact, has the standard for an
international call prefix (double zero), and the numbers in the contact
uses the '+' as international call prefix.
For e.g : 0032412345678 call and a contact using this phone number is
searched instead of +32412345678 (which is the value stored in the db).

Fine-tunning of 8c03446f57999ed23d9ee9e77797b22af9606063

OPW-2187771

closes odoo/enterprise#9347

X-original-commit: aabdb279e24be696bcd48d3b79703bdfb2c6c56f
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>