[FIX] voip: Open right partner in a phone call
Before this commit:
When we have a phone call without a partner (If we dial a number for
example), by default, we put the number in this.phone and we won't
have a mobile number. If we click on partner icon, we must jump
to the partner card, and if no partner matches this number, we must
have a new empty partner. But to search the existing partner, we will
match partner that has this phone number or partner that has no
mobile phone. So even we have no partner that has this number,
we will have the first contact that have no mobile phone.

After this commit:
When we click on partner icon, we will have the right partner
(if it exists with this phone number) or a new empty partner form.

closes odoo/enterprise#9396

X-original-commit: 23c20a9798911b6252fca3b46ce7ea9fcac1049c
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>