[IMP] hr_referral: Buy reward message improvement
Purpose of the task is, While click on 'Buy' Button, the popup is
opened with message 'Do you want to confirm this purchase? After
comfirmation, a HR will contact you.'

So this message need to be improved as: 'Do you want to confirm this
reward? After confirmation an HR will contact you.'

closes odoo/enterprise#9419

Taskid: 2205646
Closes: #9419
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>