[IMP] helpdesk: rename the alias configuration button in in helpdesk.team view
This commit rename the button that redirect to the configuration panel
where the alias configuration can be done (External email server).

The name was unclear as this is not a domain name that we can configure but
an alias instead. Behind the alias, there is a custom domain, though.

Task ID: 2214653
PR #9367
Linked to Community PR odoo/odoo#48023

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>