[FIX] web_gantt: do not edit pill on resize
Before this commit, a dialog used to open when the resizing of
a pill ended. This happened if the mouseup event was triggered
on the pill.

To fix the issue, we delay of a tick the moment when we mark the
pill as no longer being edited, s.t. when the click occurs (right
after the mouseup), the pill is still 'locked', and we do not open
the dialog to edit it.

Part of task 2202715

X-original-commit: d32eb2efff8ba3ceb0f92c26640b160911fc820d
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>