[FIX] timesheet_grid: archive root menu on uninstall
Before this commit:
    - Install `timesheet_grid`
    - Uninstall `timesheet_grid`
    -> The menu for the Timesheet app remains in the app switcher

This is confusing for the end-user as it looks like the application they
installed did not uninstall successfully and may become worried about
their billing.

This behaviour happens because the root menu for the Timesheets app is
defined inside the technical module `hr_timesheet` which is a dependence
of `timesheet_grid`, this module is not automatically uninstalled when
`timesheet_grid` is uninstalled, thus leaving the root menu untouched.

To solve this, we archive the aforementioned root menu during the
uninstall of `timesheet_grid` and we re-activate it upon its
reinstallation.

opw-2221228

closes odoo/enterprise#9596

X-original-commit: 644dcbe10890091ae8c40caf877c0fb92329c72c
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>
Signed-off-by: Adrian Torres (adt) <adt@odoo.com>