[FIX] web_dashboard: currency in multi-company
- Set your company currency (C1) to USD
- Create a new company (C2) with currency EUR
- Go to any dashboard, e.g. Purchase
- Switch from C1 to C2

The currency remains USD, i.e. the currency of C1.

This is because the currency is set at the server level based on the
default company of the user. This is not correct in a multi-currency
environment since the current company is set client-side.

We can retrive the current company from:

https://github.com/odoo/odoo/blob/b66bcfe96dfdc62a1faa26ccf9c6adfa2b8929a4/addons/web/static/src/js/chrome/abstract_web_client.js#L132

The current company is the first allowed company, as used in:

https://github.com/odoo/odoo/blob/b66bcfe96dfdc62a1faa26ccf9c6adfa2b8929a4/addons/web/static/src/js/widgets/switch_company_menu.js#L42

We add the currency per company in the session for convenience in
`companies_currency_id`, but keep `company_currency_id` for
compatibility.

opw-2224377

closes odoo/enterprise#9595

X-original-commit: 1fcbb59d8d322055f7f77ef8d5d0ace1f647d224
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>