[FIX] helpdek_timesheet: create new project from helpdesk team
Before this commit:
If you have installed helpdesk_timesheet but not
helpdesk_sale_timesheet, you will not be able to create a new team
with no project. Because the code will try to create automatically
a new project with a field that doesn't exist without
helpdesk_sale_timesheet (field: allow_billable).

After this commit:
If you have not helpdesk_sale_timesheeet, you will be able to create
a new helpdesk team with timesheet, without specifying a project.

closes odoo/enterprise#9480

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>