[IMP] web_studio: adapt media-dialog usage in test to new behaviour
In odoo/odoo#45174 the media-dialog behaviour was changed so that an
image is selected immediately when it is clicked and the dialog is not
in 'multi-image mode', as such, a test had to be adapted so that it no
longer tries to click that button. Additionally, the test tries to click
the first attachment using domUtils.click, which checks for visibility,
but that same PR makes it so that attachments are not displayed until
their image is loaded, so we also have to remove the d-none class before
trying to click.

task-2091417

closes odoo/enterprise#9520

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>