[IMP] web_studio: blockUI/unblockUI on field creation
when new field created using studio blockUI so that user can not
mess the view in between field creation, field creation takes time
so if UI is not blocked then user can mess the view while field
creation process is still not finished.

to avoid this situation blocked UI when field creation call done and
unblock UI once it is finished or failed.

task-2221093

closes odoo/enterprise#9533

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>