[IMP] account_reports: allow to use recusrive tax report declaration
Task 2214338

Lets say we have lines setup this way:
* A = B + C
* D = E + F
Before, if we wanted to do X = A + D, we had to write it like
X = B + C + E + F because we could only add lines computed from tags.

Now, we can also write it like X = A + D, which is more understandable
from a user point of view, and easier to maintain in the xml files.

closes odoo/enterprise#9534

Related: odoo/odoo#48508
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>