[FIX] hr_contract_salary: Apply several fix from refactoring
Before this commit:
1. The amount on the right for car company, was always set to 0
xhen we open the link (it's updated when we change the selected car)
2. When we fold an item (on front end), the amount of this
item (on the right) was not set to 0
3. If we set a contract template for new signature, but no contract
template signature for existing contract, we have an error.
4. If we have no hr responsible on contract, the error message
was not clear.
5. When we have no available vehicule (empty list), we can have
error in python, because we made operation on False.
6. The new vehicule to add in waiting list was never created.
7. vehicle in waitinglist have no future driver and no company.
8. When an employee validate his contract, he arrive on signature
page, then he decide to use back button, selected item are always
selected but stay folded.
9. Default car on contract has mistmatch between partner_id and
employee_id

After this commit:
No more bug from the list above