[FIX] account_sepa_direct_debit : fix error message
When posting a sdd payment, if no sdd mandate can be found an error message is triggerd.
This message included the payment name. But the name is not set at that moment. Message was :
"Unable to post payment 'False' due to no usable mandate being available at date...".
Removed name form message.

closes odoo/enterprise#9633

Task: 2228538
X-original-commit: 9a4667616fdf4000305894e5e4782acab52a6365
Signed-off-by: jbw-odoo <jbw-odoo@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>