[IMP] Field Service: improve column formatting on worksheet reports
Purpose: The user can design his worksheet templates to have 1 or 2
columns. This formatting is disregarded once the worksheet report is
generated and all the fields are put one below the other.

The purpose of this task is to keep the column formatting when the
report is generated.

Task id: 2282899
Enterprise PR: #12808

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>