[IMP] planning: implement "planning.slot" group_expand for role_id
=======
Purpose
=======

Display all roles that have been allocated in "planning.slot" in the past or future 30 days in gantt view with a group by role,
making the encoding much easier for these roles.

==============
Specifications
==============

Add a group_expand for role_id in "planning.slot".

The group_expand is applied for the following Action Window:
  - My Planning
  - Schedule > By Employee
  - Schedule > By Role

[IMP] planning: clean code

[IMP] planning: fix employee group expand after rebase