[IMP] project_forecast: implement "planning.slot" group_expand for project_id and task_id
=======
Purpose
=======

Display all projects (or tasks) that have been allocated in "planning.slot" in the past or future 30 days in gantt view with a group by project (or task),
making the encoding much easier for these projects (or tasks).

==============
Specifications
==============

Add a group_expand for project_id and task_id in "planning.slot".

The group_expand for role_id, project_id and task_id are applied for the following Action Window:
  - in Project:
    * Planning > By Employee
    * Planning > By Project

  - in Planning:
    * My Planning
    * Schedule > By Employee
    * Schedule > By Role
    * Schedule > By Project

[IMP] project_forecast: clean code

[IMP] project_forecast: fix employee group expand after rebase