[FIX] project_timesheet_forecast: use employee to compute effective hours
Before this commit:
User was used to find timesheets and the sum of hours is effective hours.
But it doesn't make sense to use the user to search timesheet as all planning
application is based on employee. And if an employee has no user, in
effective hours, we will have the sum of all timesheets that belong to
employee without user.

After this commit:
We use employee to find timesheets.

closes odoo/enterprise#9646

X-original-commit: 32085a56a7251fbddbfdbe16eb56e6126e09fd71
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: jbm-odoo <jbm-odoo@users.noreply.github.com>