[FIX] mrp_mps: MPS with Subcontracted products
- Install MPS and activate subcontracting
- Create a Subcontracted product P, set a Subcontracted BOM and a
  supplier
- Add P to the MPS
- Show 'Actual Replenishment'
- Add some quantity to the first week, Replenish
- Validate the PO

The 'Actual Replenishment' doesn't show the quantity validated on the
PO.

This happens because the `stock.move` domain doesn't take into account
internal locations, only vendor locations. Since the subcontracting
location is an internal location, the quantity is not counted.

We take into account internal locations if they are not in the
warehouse. This way, it takes into account subcontracting. Note that
resupplies from another warehouse use `transit` location usage, so they
are not impacted. However, direct transfers between 2 warehouses are now
taken into account.

opw-2220850

closes odoo/enterprise#9664

X-original-commit: ad05436e2a0fa0423a5014f1cffdb4f87180f9bf
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>