[FIX] industry_fsm_sale: Fix the traceback on installation.
Currently, during installation of industry_fsm_sale there is trackback
for button 'action_view_invoices' which actually does not exist in
parent view. Button is added from the industry_fsm_sale by
inheritance so it will work when directly install the indurtry_fsm_sale
due to priority of the view but does not work when you install
sales > project > field service.

So change the xpath of button to add button inside the button_box.

closes odoo/enterprise#9819

Taskid: 2230174
Closes: #9651
X-original-commit: fd9f0623d75bfb2a2d577c654f42d0e111b82b19
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>